+++ src/config/jest-setup.ts
@@ -9,6 +9,5 @@ jest.mock('firebase/auth', () => ({
     authDomain: '',
     appName: '',
   }),
-  // eslint-disable-next-line @typescript-eslint/no-empty-function
-  onAuthStateChanged: () => () => {},
+  onAuthStateChanged: () => jest.fn(),
 }));
