+++ src/screens/Auth/Welcome/Welcome.spec.tsx
