+++ src/screens/Auth/Welcome/Welcome.styles.tsx
