+++ src/screens/Auth/Welcome/Welcome.tsx
