+++ src/screens/Auth/Welcome/index.ts
