+++ src/screens/index.ts
@@ -1,6 +1,6 @@
 export * from './Home';
 export * from './Menu';
-export * from './Welcome';
-export * from './Login';
-export * from './Register';
-export * from './ForgotPassword';
+export * from './Auth/Welcome';
+export * from './Auth/Login';
+export * from './Auth/Register';
+export * from './Auth/ForgotPassword';
